+++ src/main/kotlin/app/revanced/patches/music/player/components/fingerprints/PlayerViewPagerConstructorFingerprint.kt
@@ -1,0 +1,17 @@
+package app.revanced.patches.music.player.components.fingerprints
+
+import app.revanced.patcher.extensions.or
+import app.revanced.patcher.fingerprint.MethodFingerprint
+import app.revanced.patches.music.utils.resourceid.SharedResourceIdPatch.MiniPlayerViewPager
+import app.revanced.patches.music.utils.resourceid.SharedResourceIdPatch.PlayerViewPager
+import app.revanced.util.containsWideLiteralInstructionIndex
+import com.android.tools.smali.dexlib2.AccessFlags
+
+internal object PlayerViewPagerConstructorFingerprint : MethodFingerprint(
+    returnType = "V",
+    accessFlags = AccessFlags.PUBLIC or AccessFlags.CONSTRUCTOR,
+    customFingerprint = { methodDef, _ ->
+        methodDef.containsWideLiteralInstructionIndex(MiniPlayerViewPager)
+                && methodDef.containsWideLiteralInstructionIndex(PlayerViewPager)
+    },
+)