+++ src/main/kotlin/app/revanced/patches/shared/viewgroup/fingerprints/ViewGroupMarginFingerprint.kt
@@ -1,0 +1,11 @@
+package app.revanced.patches.shared.viewgroup.fingerprints
+
+import app.revanced.patcher.extensions.or
+import app.revanced.patcher.fingerprint.MethodFingerprint
+import com.android.tools.smali.dexlib2.AccessFlags
+
+internal object ViewGroupMarginFingerprint : MethodFingerprint(
+    returnType = "V",
+    accessFlags = AccessFlags.PUBLIC or AccessFlags.STATIC,
+    parameters = listOf("Landroid/view/View;", "I", "I"),
+)