+++ src/main/kotlin/app/revanced/patches/youtube/layout/branding/icon/CustomBrandingIconPatch.kt
@@ -22,7 +22,7 @@ object CustomBrandingIconPatch : BaseResourcePatch(
     compatiblePackages = COMPATIBLE_PACKAGE,
     use = false,
 ) {
-    private const val DEFAULT_ICON_KEY = "Revancify Blue"
+    private const val DEFAULT_ICON_KEY = "Xisr Yellow"
 
     private val availableIcon = mapOf(
         "AFN Blue" to "afn_blue",
@@ -31,11 +31,11 @@ object CustomBrandingIconPatch : BaseResourcePatch(
         "MMT Blue" to "mmt_blue",
         "MMT Green" to "mmt_green",
         "MMT Yellow" to "mmt_yellow",
-        DEFAULT_ICON_KEY to "revancify_blue",
+        "Revancify Blue" to "revancify_blue",
         "Revancify Red" to "revancify_red",
-        "Revancify Yellow" to "revancify_yellow",
         "Vanced Black" to "vanced_black",
         "Vanced Light" to "vanced_light",
+        DEFAULT_ICON_KEY to "xisr_yellow",
         "YouTube" to "youtube"
     )
 
@@ -179,7 +179,7 @@ object CustomBrandingIconPatch : BaseResourcePatch(
         key = "RestoreOldSplashAnimation",
         default = true,
         title = "Restore old splash animation",
-        description = "Restores old style splash animation.",
+        description = "Restore the old style splash animation.",
         required = true
     )
 
@@ -192,7 +192,6 @@ object CustomBrandingIconPatch : BaseResourcePatch(
             .underBarOrThrow()
 
         val appIconResourcePath = "youtube/branding/$appIcon"
-        val stockResourcePath = "youtube/branding/stock"
 
         // Check if a custom path is used in the patch options.
         if (!availableIcon.containsValue(appIcon)) {
@@ -260,7 +259,7 @@ object CustomBrandingIconPatch : BaseResourcePatch(
                 }
 
                 context.copyXmlNode(
-                    "$stockResourcePath/splash",
+                    "$appIconResourcePath/splash",
                     "values-v31/styles.xml",
                     "resources"
                 )