+++ src/main/kotlin/app/revanced/patches/youtube/utils/resourceid/SharedResourceIdPatch.kt
@@ -17,6 +17,7 @@ import app.revanced.patches.shared.mapping.ResourceType.STYLE
 
 @Patch(dependencies = [ResourceMappingPatch::class])
 object SharedResourceIdPatch : ResourcePatch() {
+    var AccessibilityOfflineButtonSync = -1L
     var AccountSwitcherAccessibility = -1L
     var ActionBarRingo = -1L
     var ActionBarRingoBackground = -1L
@@ -110,6 +111,7 @@ object SharedResourceIdPatch : ResourcePatch() {
     var VideoQualityBottomSheet = -1L
     var VarispeedUnavailableTitle = -1L
     var VideoQualityUnavailableAnnouncement = -1L
+    var VideoZoomSnapIndicator = -1L
     var VoiceSearch = -1L
     var YouTubeControlsOverlaySubtitleButton = -1L
     var YtOutlinePictureInPictureWhite = -1L
@@ -120,6 +122,7 @@ object SharedResourceIdPatch : ResourcePatch() {
 
     override fun execute(context: ResourceContext) {
 
+        AccessibilityOfflineButtonSync = getId(STRING, "accessibility_offline_button_sync")
         AccountSwitcherAccessibility = getId(STRING, "account_switcher_accessibility_label")
         ActionBarRingo = getId(LAYOUT, "action_bar_ringo")
         ActionBarRingoBackground = getId(LAYOUT, "action_bar_ringo_background")
@@ -218,6 +221,7 @@ object SharedResourceIdPatch : ResourcePatch() {
         VarispeedUnavailableTitle = getId(STRING, "varispeed_unavailable_title")
         VideoQualityUnavailableAnnouncement =
             getId(STRING, "video_quality_unavailable_announcement")
+        VideoZoomSnapIndicator = getId(ID, "video_zoom_snap_indicator")
         VoiceSearch = getId(ID, "voice_search")
         YouTubeControlsOverlaySubtitleButton =
             getId(LAYOUT, "youtube_controls_overlay_subtitle_button")